+++ components/dashboard/src/teams/TeamPolicies.tsx
@@ -6,7 +6,7 @@
 
 import { isGitpodIo } from "../utils";
 import { OrganizationSettings } from "@gitpod/public-api/lib/gitpod/v1/organization_pb";
-import { FormEvent, useCallback, useEffect, useMemo, useState } from "react";
+import { FormEvent, useCallback, useEffect, useState } from "react";
 import Alert from "../components/Alert";
 import { CheckboxInputField } from "../components/forms/CheckboxInputField";
 import { Heading2, Heading3, Subheading } from "../components/typography/headings";
@@ -15,20 +15,8 @@ import { useOrgSettingsQuery } from "../data/organizations/org-settings-query";
 import { useCurrentOrg } from "../data/organizations/orgs-query";
 import { useUpdateOrgSettingsMutation } from "../data/organizations/update-org-settings-mutation";
 import { OrgSettingsPage } from "./OrgSettingsPage";
-import { Button } from "@podkit/buttons/Button";
-import { useAllowedWorkspaceClassesMemo } from "../data/workspaces/workspace-classes-query";
 import { ConfigurationSettingsField } from "../repositories/detail/ConfigurationSettingsField";
-import {
-    WorkspaceClassesModifyModal,
-    WorkspaceClassesModifyModalProps,
-    WorkspaceClassesOptions,
-} from "../components/WorkspaceClassesOptions";
-import { useMutation } from "@tanstack/react-query";
-import { useAllowedWorkspaceEditorsMemo } from "../data/ide-options/ide-options-query";
-import { IdeOptions, IdeOptionsModifyModal, IdeOptionsModifyModalProps } from "../components/IdeOptions";
 import { useDocumentTitle } from "../hooks/use-document-title";
-import { LinkButton } from "@podkit/buttons/LinkButton";
-import PillLabel from "../components/PillLabel";
 import { useOrgBillingMode } from "../data/billing-mode/org-billing-mode-query";
 import { converter } from "../service/public-api";
 import { useToast } from "../components/toasts/Toasts";
@@ -38,13 +26,11 @@ import { Link } from "react-router-dom";
 import { InputField } from "../components/forms/InputField";
 import { TextInput } from "../components/forms/TextInputField";
 import { LoadingButton } from "@podkit/buttons/LoadingButton";
-import {
-    OrganizationRoleRestrictionModal,
-    OrganizationRoleRestrictionModalProps,
-    OrgMemberPermissionRestrictionsOptions,
-} from "../components/OrgMemberPermissionsOptions";
-import { LightbulbIcon } from "lucide-react";
 import { MaxParallelWorkspaces } from "./policies/MaxParallelWorkspaces";
+import { WorkspaceClassesEnterpriseCallout } from "./policies/WorkspaceClassesEnterpriseCallout";
+import { EditorOptions } from "./EditorOptions";
+import { RolePermissionsRestrictions } from "./policies/RoleRestrictions";
+import { OrgWorkspaceClassesOptions } from "./policies/OrgWorkspaceClassesOptions";
 
 export default function TeamPoliciesPage() {
     useDocumentTitle("Organization Settings - Policies");
@@ -240,262 +226,3 @@ export default function TeamPoliciesPage() {
         </>
     );
 }
-
-interface OrgWorkspaceClassesOptionsProps {
-    isOwner: boolean;
-    settings?: OrganizationSettings;
-    handleUpdateTeamSettings: (
-        newSettings: Partial<OrganizationSettings>,
-        options?: { throwMutateError?: boolean },
-    ) => Promise<void>;
-}
-const OrgWorkspaceClassesOptions = ({
-    isOwner,
-    settings,
-    handleUpdateTeamSettings,
-}: OrgWorkspaceClassesOptionsProps) => {
-    const [showModal, setShowModal] = useState(false);
-    const { data: allowedClassesInOrganization, isLoading: isLoadingClsInOrg } = useAllowedWorkspaceClassesMemo(
-        undefined,
-        {
-            filterOutDisabled: true,
-            ignoreScope: ["configuration"],
-        },
-    );
-    const { data: allowedClassesInInstallation, isLoading: isLoadingClsInInstall } = useAllowedWorkspaceClassesMemo(
-        undefined,
-        {
-            filterOutDisabled: true,
-            ignoreScope: ["organization", "configuration"],
-        },
-    );
-
-    const restrictedWorkspaceClasses = useMemo(() => {
-        const allowedList = settings?.allowedWorkspaceClasses ?? [];
-        if (allowedList.length === 0) {
-            return [];
-        }
-        return allowedClassesInInstallation.filter((cls) => !allowedList.includes(cls.id)).map((cls) => cls.id);
-    }, [settings?.allowedWorkspaceClasses, allowedClassesInInstallation]);
-
-    const updateMutation: WorkspaceClassesModifyModalProps["updateMutation"] = useMutation({
-        mutationFn: async ({ restrictedWorkspaceClasses }) => {
-            let allowedWorkspaceClasses = allowedClassesInInstallation.map((e) => e.id);
-            if (restrictedWorkspaceClasses.length > 0) {
-                allowedWorkspaceClasses = allowedWorkspaceClasses.filter(
-                    (e) => !restrictedWorkspaceClasses.includes(e),
-                );
-            }
-            const allAllowed = allowedClassesInInstallation.every((e) => allowedWorkspaceClasses.includes(e.id));
-            if (allAllowed) {
-                // empty means allow all classes
-                allowedWorkspaceClasses = [];
-            }
-            await handleUpdateTeamSettings({ allowedWorkspaceClasses }, { throwMutateError: true });
-        },
-    });
-
-    return (
-        <ConfigurationSettingsField>
-            <Heading3>Available workspace classes</Heading3>
-            <Subheading>
-                Limit the available workspace classes in your organization. Requires{" "}
-                <span className="font-medium">Owner</span> permissions to change.
-            </Subheading>
-
-            <WorkspaceClassesOptions
-                isLoading={isLoadingClsInOrg}
-                className="mt-4"
-                classes={allowedClassesInOrganization}
-            />
-
-            {isOwner && (
-                <Button className="mt-6" onClick={() => setShowModal(true)}>
-                    Manage Classes
-                </Button>
-            )}
-
-            {showModal && (
-                <WorkspaceClassesModifyModal
-                    isLoading={isLoadingClsInInstall}
-                    showSetDefaultButton={false}
-                    showSwitchTitle={false}
-                    restrictedWorkspaceClasses={restrictedWorkspaceClasses}
-                    allowedClasses={allowedClassesInInstallation}
-                    updateMutation={updateMutation}
-                    onClose={() => setShowModal(false)}
-                />
-            )}
-        </ConfigurationSettingsField>
-    );
-};
-
-type RolePermissionsRestrictionsProps = {
-    settings: OrganizationSettings | undefined;
-    isOwner: boolean;
-    handleUpdateTeamSettings: (
-        newSettings: Partial<PlainMessage<OrganizationSettings>>,
-        options?: { throwMutateError?: boolean },
-    ) => Promise<void>;
-};
-
-const RolePermissionsRestrictions = ({
-    settings,
-    isOwner,
-    handleUpdateTeamSettings,
-}: RolePermissionsRestrictionsProps) => {
-    const [showModal, setShowModal] = useState(false);
-
-    const updateMutation: OrganizationRoleRestrictionModalProps["updateMutation"] = useMutation({
-        mutationFn: async ({ roleRestrictions }) => {
-            await handleUpdateTeamSettings({ roleRestrictions }, { throwMutateError: true });
-        },
-    });
-
-    return (
-        <ConfigurationSettingsField>
-            <Heading3>Roles allowed to start workspaces from non-imported repos</Heading3>
-            <Subheading className="mb-2">
-                Restrict specific roles from initiating workspaces using non-imported repositories. This setting
-                requires <span className="font-medium">Owner</span> permissions to modify.
-                <br />
-                <span className="flex flex-row items-center gap-1 my-2">
-                    <LightbulbIcon size={20} />{" "}
-                    <span>
-                        Tip: Imported repositories are those listed under{" "}
-                        <Link to={"/repositories"} className="gp-link">
-                            Repository settings
-                        </Link>
-                        .
-                    </span>
-                </span>
-            </Subheading>
-
-            <OrgMemberPermissionRestrictionsOptions roleRestrictions={settings?.roleRestrictions ?? []} />
-
-            {isOwner && (
-                <Button className="mt-6" onClick={() => setShowModal(true)}>
-                    Manage Permissions
-                </Button>
-            )}
-
-            {showModal && (
-                <OrganizationRoleRestrictionModal
-                    isLoading={false}
-                    defaultClass={""}
-                    roleRestrictions={settings?.roleRestrictions ?? []}
-                    showSetDefaultButton={false}
-                    showSwitchTitle={false}
-                    allowedClasses={[]}
-                    updateMutation={updateMutation}
-                    onClose={() => setShowModal(false)}
-                />
-            )}
-        </ConfigurationSettingsField>
-    );
-};
-
-interface EditorOptionsProps {
-    settings: OrganizationSettings | undefined;
-    isOwner: boolean;
-    handleUpdateTeamSettings: (
-        newSettings: Partial<OrganizationSettings>,
-        options?: { throwMutateError?: boolean },
-    ) => Promise<void>;
-}
-const EditorOptions = ({ isOwner, settings, handleUpdateTeamSettings }: EditorOptionsProps) => {
-    const [showModal, setShowModal] = useState(false);
-    const { data: installationOptions, isLoading: installationOptionsIsLoading } = useAllowedWorkspaceEditorsMemo(
-        undefined,
-        {
-            filterOutDisabled: true,
-            ignoreScope: ["organization", "configuration"],
-        },
-    );
-    const { data: orgOptions, isLoading: orgOptionsIsLoading } = useAllowedWorkspaceEditorsMemo(undefined, {
-        filterOutDisabled: true,
-        ignoreScope: ["configuration"],
-    });
-
-    const updateMutation: IdeOptionsModifyModalProps["updateMutation"] = useMutation({
-        mutationFn: async ({ restrictedEditors, pinnedEditorVersions }) => {
-            const updatedRestrictedEditors = [...restrictedEditors.keys()];
-            const updatedPinnedEditorVersions = Object.fromEntries(pinnedEditorVersions.entries());
-
-            await handleUpdateTeamSettings(
-                { restrictedEditorNames: updatedRestrictedEditors, pinnedEditorVersions: updatedPinnedEditorVersions },
-                { throwMutateError: true },
-            );
-        },
-    });
-
-    const restrictedEditors = new Set<string>(settings?.restrictedEditorNames || []);
-    const pinnedEditorVersions = new Map<string, string>(Object.entries(settings?.pinnedEditorVersions || {}));
-
-    return (
-        <ConfigurationSettingsField>
-            <Heading3>Available editors</Heading3>
-            <Subheading>
-                Limit the available editors in your organization. Requires <span className="font-medium">Owner</span>{" "}
-                permissions to change.
-            </Subheading>
-
-            <IdeOptions
-                isLoading={orgOptionsIsLoading}
-                className="mt-4"
-                ideOptions={orgOptions}
-                pinnedEditorVersions={pinnedEditorVersions}
-            />
-
-            {isOwner && (
-                <Button className="mt-6" onClick={() => setShowModal(true)}>
-                    Manage Editors
-                </Button>
-            )}
-
-            {showModal && (
-                <IdeOptionsModifyModal
-                    isLoading={installationOptionsIsLoading}
-                    ideOptions={installationOptions}
-                    restrictedEditors={restrictedEditors}
-                    pinnedEditorVersions={pinnedEditorVersions}
-                    updateMutation={updateMutation}
-                    onClose={() => setShowModal(false)}
-                />
-            )}
-        </ConfigurationSettingsField>
-    );
-};
-
-const WorkspaceClassesEnterpriseCallout = () => {
-    return (
-        <ConfigurationSettingsField className="bg-pk-surface-secondary">
-            <Heading3 className="flex items-center gap-4">
-                Additional workspace classes
-                <PillLabel type="warn">Enterprise</PillLabel>
-            </Heading3>
-            <Subheading>
-                Access to more powerful workspace classes with up to 30 cores 54GB of RAM and 100GB of storage
-            </Subheading>
-
-            <div className="mt-6 flex flex-row space-x-2">
-                <LinkButton
-                    variant="secondary"
-                    className="border border-pk-content-tertiary text-pk-content-primary bg-pk-surface-primary"
-                    href="https://www.gitpod.io/docs/configure/workspaces/workspace-classes#enterprise"
-                    isExternalUrl={true}
-                >
-                    Documentation
-                </LinkButton>
-                <LinkButton
-                    variant="secondary"
-                    className="border border-pk-content-tertiary text-pk-content-primary bg-pk-surface-primary"
-                    href="https://www.gitpod.io/docs/enterprise"
-                    isExternalUrl={true}
-                >
-                    Learn more about Enterprise
-                </LinkButton>
-            </div>
-        </ConfigurationSettingsField>
-    );
-};
